api/requirements.txt
flask
flask-cors
supabase
anthropic
pypdf2
gunicorn
python-dotenv
requests
pyjwt[crypto]